+++ dlint/linters/__init__.py
@@ -31,6 +31,7 @@ from .bad_sys_use import BadSysUseLinter
 from .bad_tarfile_use import BadTarfileUseLinter
 from .bad_tempfile_use import BadTempfileUseLinter
 from .bad_urllib3_module_attribute_use import BadUrllib3ModuleAttributeUseLinter
+from .bad_urllib3_kwarg_use import BadUrllib3KwargUseLinter
 from .bad_pickle_use import BadPickleUseLinter
 from .bad_xml_use import BadXMLUseLinter
 from .bad_xmlrpc_use import BadXmlrpcUseLinter
@@ -64,6 +65,7 @@ ALL = (
     BadTempfileUseLinter,
     BadTarfileUseLinter,
     BadUrllib3ModuleAttributeUseLinter,
+    BadUrllib3KwargUseLinter,
     BadPickleUseLinter,
     BadXMLUseLinter,
     BadXmlrpcUseLinter,